move full_like to fluid.layers
