+++ python/paddle/fluid/tests/unittests/test_fill_any_like_op.py
@@ -106,8 +106,9 @@ class TestFillAnyLikeOp_attr_out(unittest.TestCase):
         with fluid.program_guard(train_program, startup_program):
             fill_value = 2.0
             input = fluid.data(name='input', dtype='float32', shape=[2, 3])
-            output = paddle.full_like(input, fill_value)
-            output_dtype = paddle.full_like(input, fill_value, dtype='float32')
+            output = fluid.layers.full_like(input, fill_value)
+            output_dtype = fluid.layers.full_like(
+                input, fill_value, dtype='float32')
 
             place = fluid.CPUPlace()
             if fluid.core.is_compiled_with_cuda():
@@ -133,20 +134,20 @@ class TestFillAnyLikeOpError(unittest.TestCase):
             #for ci coverage
 
             input_data = fluid.data(name='input', dtype='float32', shape=[2, 3])
-            output = paddle.full_like(input_data, 2.0)
+            output = fluid.layers.full_like(input_data, 2.0)
 
             def test_input_dtype():
-                paddle.full_like
+                fluid.layers.full_like
 
             self.assertRaises(
                 ValueError,
-                paddle.full_like,
+                fluid.layers.full_like,
                 input=input_data,
                 fill_value=2,
                 dtype='uint4')
             self.assertRaises(
                 TypeError,
-                paddle.full_like,
+                fluid.layers.full_like,
                 input=input_data,
                 fill_value=2,
                 dtype='int16')
